Refactored ufunc_per_section and moved ParameterIndex_Ended to datastore_utils
